+++ src/main/java/com/xiliulou/electricity/controller/admin/ElectricityCabinetOrderAdminController.java
@@ -1,4 +1,5 @@
 package com.xiliulou.electricity.controller.admin;
+
 import com.xiliulou.core.web.R;
 import com.xiliulou.electricity.query.ElectricityCabinetOrderQuery;
 import com.xiliulou.electricity.service.ElectricityCabinetOrderService;
@@ -28,8 +29,8 @@ public class ElectricityCabinetOrderAdminController {
 
     //换电柜订单查询
     @GetMapping("/admin/electricityCabinetOrder/list")
-    public R queryList(@RequestParam(value = "size", required = false) Integer size,
-                       @RequestParam(value = "offset", required = false) Integer offset,
+    public R queryList(@RequestParam(value = "size", required = false) Long size,
+                       @RequestParam(value = "offset", required = false) Long offset,
                        @RequestParam(value = "orderId", required = false) String orderId,
                        @RequestParam(value = "phone", required = false) String phone,
                        @RequestParam(value = "areaId", required = false) Integer areaId,
@@ -39,11 +40,11 @@ public class ElectricityCabinetOrderAdminController {
                        @RequestParam(value = "paymentMethod", required = false) Integer paymentMethod) {
 
         if (Objects.isNull(size)) {
-            size = 10;
+            size = 10L;
         }
 
         if (Objects.isNull(offset) || offset < 0) {
-            offset = 0;
+            offset = 0L;
         }
 
         ElectricityCabinetOrderQuery electricityCabinetOrderQuery = ElectricityCabinetOrderQuery.builder()
@@ -70,15 +71,15 @@ public class ElectricityCabinetOrderAdminController {
     //换电柜订单导出报表
     @GetMapping("/admin/electricityCabinetOrder/exportExcel")
     public void exportExcel(@RequestParam(value = "orderId", required = false) String orderId,
-                       @RequestParam(value = "phone", required = false) String phone,
-                       @RequestParam(value = "areaId", required = false) Integer areaId,
-                       @RequestParam(value = "status", required = false) Integer status,
-                       @RequestParam(value = "beginTime", required = false) Long beginTime,
-                       @RequestParam(value = "endTime", required = false) Long endTime, HttpServletResponse response) {
+                            @RequestParam(value = "phone", required = false) String phone,
+                            @RequestParam(value = "areaId", required = false) Integer areaId,
+                            @RequestParam(value = "status", required = false) Integer status,
+                            @RequestParam(value = "beginTime", required = false) Long beginTime,
+                            @RequestParam(value = "endTime", required = false) Long endTime, HttpServletResponse response) {
 
-        Double days=(Double.valueOf(endTime-beginTime))/1000/3600/24;
-        if(days>31){
-           return;
+        Double days = (Double.valueOf(endTime - beginTime)) / 1000 / 3600 / 24;
+        if (days > 31) {
+            return;
         }
         ElectricityCabinetOrderQuery electricityCabinetOrderQuery = ElectricityCabinetOrderQuery.builder()
                 .orderId(orderId)
@@ -87,7 +88,7 @@ public class ElectricityCabinetOrderAdminController {
                 .status(status)
                 .beginTime(beginTime)
                 .endTime(endTime).build();
-       electricityCabinetOrderService.exportExcel(electricityCabinetOrderQuery,response);
+        electricityCabinetOrderService.exportExcel(electricityCabinetOrderQuery, response);
     }
