+++ src/main/java/com/xiliulou/electricity/mapper/UserMapper.java
@@ -65,5 +65,5 @@ public interface UserMapper extends BaseMapper<User> {
      */
     int deleteById(Long uid);
 
-    IPage queryListUserByCriteria(Page page, @Param("uid") Long uid, @Param("size") Integer size, @Param("offset") Integer offset, @Param("name") String name, @Param("phone") String phone, @Param("type") Integer type, @Param("startTime") Long startTime, @Param("endTime") Long endTime);
+    IPage queryListUserByCriteria(Page page, @Param("uid") Long uid, @Param("size") Long size, @Param("offset") Long offset, @Param("name") String name, @Param("phone") String phone, @Param("type") Integer type, @Param("startTime") Long startTime, @Param("endTime") Long endTime);
 }
