+++ src/main/java/com/xiliulou/electricity/query/ElectricityCabinetQuery.java
@@ -1,4 +1,5 @@
 package com.xiliulou.electricity.query;
+
 import lombok.Builder;
 import lombok.Data;
 
@@ -10,40 +11,40 @@ import lombok.Data;
 @Data
 @Builder
 public class ElectricityCabinetQuery {
-	private Integer size;
-	private Integer offset;
-	/**
-	 * 电池编号
-	 */
-	private String sn;
-	/**
-	 * 换电柜名称
-	 */
-	private String name;
-	/**
-	 * 换电柜地区Id
-	 */
-	private Integer areaId;
-	/**
-	 * 换电柜地址
-	 */
-	private String address;
-	/**
-	 * 可用状态(0--启用，1--禁用)
-	 */
-	private Integer usableStatus;
-	/**
-	 * 电源状态(0--通电，1--断电)
-	 */
-	private Integer powerStatus;
-	/**
-	 * 物联网连接状态（0--连网，1--断网）
-	 */
-	private Integer onlineStatus;
-	private Double distance;
-	private Double lon;
-	private Double lat;
+    private Long size;
+    private Long offset;
+    /**
+     * 电池编号
+     */
+    private String sn;
+    /**
+     * 换电柜名称
+     */
+    private String name;
+    /**
+     * 换电柜地区Id
+     */
+    private Integer areaId;
+    /**
+     * 换电柜地址
+     */
+    private String address;
+    /**
+     * 可用状态(0--启用，1--禁用)
+     */
+    private Integer usableStatus;
+    /**
+     * 电源状态(0--通电，1--断电)
+     */
+    private Integer powerStatus;
+    /**
+     * 物联网连接状态（0--连网，1--断网）
+     */
+    private Integer onlineStatus;
+    private Double distance;
+    private Double lon;
+    private Double lat;
 
-	private Long beginTime;
-	private Long endTime;
+    private Long beginTime;
+    private Long endTime;
 }
